Compactor should account for checkpoints when merging manifest state (#437)
Similar to #435. The compactor also needs to account for newly added
checkpoints.

Worth noting that @rodesai and I have been discussing a more robust
solution in which the writer and compactor can record their changes in
some kind of Delta struct and then apply them on top of the latest
state. I've opened https://github.com/slatedb/slatedb/issues/436 for
this potential improvement.