Replace `Bytes` with `&[u8]` and `Vec<u8>` in `MergeOperator`
As a follow-on to #450, I'm eliminating `Bytes` from user-facing APIs. The
`MergeOperator` currently uses `Bytes` for inputs and outputs. I've replaced the
inputs with `&[u8]` and the output with `Vec<u8>`. I opted against `AsRef<[u8]>`
for the inputs since this is a callback from our own SlateDB code, and we know
what we're passing. We don't need the added flexibility. Plus, it broke the
vtable lookups to include generics in a trait. I opted for `Vec<u8>` for the
return in keeping with our other public return methods.

I took a look at [`Bytes::from` for `Vec<u8>`](https://docs.rs/bytes/latest/src/bytes/bytes.rs.html#964-997).
It seems not to do unnecessary copies, so I think it should be OK for us here.