Rename `bytes.rs` and re-export `bytes` crate (#457)
This PR renames `bytes.rs` to `bytes_generator.rs` so that we can
re-export the `bytes` crate without a name conflict. This is to lay the
groundwork for #450, where we've decided we want to keep `Bytes` as the
return type for our API. To make life easier on users, I think we should
re-export the `bytes` crate since we're using it in our public API (like
we do with the `object_store` and `fail_parallel` crates).

I considered deleting `bytes.rs` and moving its content into
`test_utils.rs`, but `OrderedBytesGenerator` is used by the `bencher`
feature, which is a feature flag, not a test. When importing
`test_utils` for `bencher`, it was causing a bunch of "unused" errors
because of all the test functions that weren't used.

I also considered trying to use `extern` to somehow alias the `bytes`
crate in `lib.rs`, but was unsuccessful. I don't want to export it under
a separate name because that would be confusing to users.

This led me to simply rename the `bytes.rs` file to get around the
conflict.