restructure exports to top level of crate (#463)
This patch makes a couple of changes to the way public names are
exported:
- It exports many of the public names at the top level of the crate
(e.g. lib.rs) to avoid leaking the internal module structure. I retained
a few public modules for cases where I think it makes sense to expose
the contained names within a module because the components should be
grouped (e.g. configs, metrics, pluggable stuff like the dbcache)
- It adds non_exhaustive to all enums to force any users matching on the
enums to include wildcards in case new variants are added.
- Adds non_exhaustive to all pub structs with all pub fields that are
constructed by calling into a fn exported by slatedb to prevent users
from constructing these directly. In the case that the struct contains a
private or pub(crate) field we don't add non_exhaustive as these cannot
be created by users anyway. In cases where the struct implements Default
we don't add non_exhaustive to retain the ability to construct the type
using the functional update syntax with the result of `default` as this
is a better way to ensure compatibility than adding new constructor fns
whenever we add fields.