fix bug with cloning retry that uses different checkpoint
we had a bug in our clone logic where:
1. cloning fails during copying of the WAL
2. a retry of the clone operation uses a different checkpoint

In this case, the WALs written in (1) are still present, so when we go
to write to the cloned db, the write will fail with a CAS error. To fix
this problem, we just fail the clone if we see that the checkpoint has
been established but it points to a different manifest from the requested
checkpoint.