Use `u64` for SST metadata offset to support big SSTs (#484)
Arrow's `ObjectStore` crate has changed their `Range` and `usize`
lengths to `u64`. This will break SlateDB when we upgrade to the latest
version.

While investigating what needs to change on our end, I noticed that we
use a `u32` for our metadata offset in SSTs. This offset is at the tail
end of each SST file and describes where to seek to in order to get
index and filter data.

Using a `u32` for the metadata offset means we can't have an SST larger
than ~4GiB. After some discussion on Discord, we decided it's worth
supporting larger SSTs as well. To make this change, I've:

- Changed the SST's current_len to be `u64`
- Updated `read_info` to read 8 bytes for the metadata offset
- Updated filter and index offsets to use `u64` throughout
- Updated `ReadOnlyBlob` to use u64 for read range and length