support more dynamic creation/read of statistics (#482)
This patch aims to solve a few problems with the way we export stats:
- Its very tedious to export statistics. For every new statistic you
have to write some new code to read the statistic value and export it to
where you want to query/visualize your metrics.
- Statistics are all statically configured at compile time. At the same
time, SlateDB has a few components that can be plugged in at runtime by
implementing some trait. A good example now is the block cache. I expect
other similar examples will emerge, like the compaction policy. There's
no great way to expose statistics from these components, other than to
have the user get to them by holding a reference to the component and
querying them directly.
- All stats are dumped into a single giant stats struct. Though not a
huge problem, its not ideal for a couple reasons. First, it would be
better for statistics not to be writable outside of some scope. For
example, the compactor shouldn't able to be able to touch os cache
statistics. It also makes the stats a bit harder to go through since
they need to be manually grouped.

This patch tries to solve these problems by doing the following:

First, stats are exposed to the user via a new StatRegistry struct.
Stats can be added to the registry by name, looked up by name, and
listed. This way, users that want to exfiltrate all stats can simply
list the registry and send all observed stats. Components can register
their own stats at runtime. To avoid collisions, we use a name prefixed
by the name of the component writing the statistic followed by "/".

Second, each component maintains its stats separately, and adds them to
the registry. This provides some isolation between the stats of
different components.

At this point I'm looking for feedback on whether folks think the basic
problems and solution make sense. So it's probably premature to do a
detailed code review. If folks are on-board with the direction, I'll
clean it up and take it out of draft mode.

Co-authored-by: Chris Riccomini <criccomini@users.noreply.github.com>