fix: cleanup compaction job on failure (#486)
This patch makes a few changes to ensure we cleanup compaction jobs when
they fail:
- We now track compaction jobs by a unique id rather than the
destination. Though this is not strictly required, it feels a bit safer
as a way to correlate msgs from the executor with the compactor's state.
- When a compaction fails, the executor's message includes both the id
and the error.
- The compactor orchestrator clears the compaction job from its state
when a compaction fails.
- Changes the executor to use a monitored background task for
compaction, and to send the orchestrator the result from the task's
cleanup fn. This way we also report failures due to a compaction
panic'ing. This also required changing the cleanup fn to take a Result
rather than just an error, and updating the existing cleanup fns (e.g.
for flush, batch-write, etc) accordingly.

Finally, to make this change easier to test this patch refactors the
CompactorOrchestrator by splitting it into:
- CompactorEventHandler has all the logic for handling events form the
orchestrator's event loop.
- CompactorOrchestrator now just runs the event loop and delegates
handling to CompactorEventHandler. This makes it easier to test the
orchestrator's event handling because the test can generate a
deterministic sequence of events to handle.

The tests for the orchestrator have been refactored a bit to move all
the common setup into CompactorEventHandlerTestFixture.