fix: clamp the buffers backing entries inserted to the block cache
This patch changes the cache wrapper to clamp buffers backing entries
to the block cache to the minimally sized buffer required.

Normally, the buffer backing an entry may be significantly larger than the
entry itself:
- some entries (like data blocks) internally contain slices that point
  to a larger buffer that was used to range a larger chunk of data (e.g.
  in the data block example, a read range will read a range of blocks
  and hand each block a slice)
- even entries derived from reads of just the relevant data (like indexes
  and filters) may contain larger backing buffers due to the behaviour of
  dynamically sized buffer data structures like Vec, which allocate the
  data by doubling it when required to achieve amortized O(1) updates.

Consequently, the block cache may think that it's cached some number of
bytes when in reality the associated memory is much larger. This can make
it difficult/impossible to set a reasonable limit on the cache size,
leading to OOM. It also wastes memory.

This patch fixes this problem by reallocating the buffer backing cache
entries to a minimally sized buffer and copying the data. This does,
represent a performance tradeoff but it should be relatively small
compared to the existing cost of getting a cache entry from the backing
store before caching it.