add stats about block cache hits/misses and usage (#491)
This patch adds some stats about block cache accesses and usage. The
stats are measured and reported in a way that works when the cache is
shared between multiple slatedb instances:
- The stats are added to a separate struct (DbCacheStats) from the main
DbStats struct.
- The stats are measured via a wrapper of the main cache backend arc
called DbCacheWrapper. This patch changes the type that's passed into
the config and stored in the TableStore to DbCacheWrapper. Users now
need to create a DbCacheWrapper instance from the Arc<dyn DbCache> if
they want to provide a specific cache implementation.
- The stats can be accessed via the stats fn in DbCacheWrapper. So users
that pass a single DbCacheWrapper instance to multiple SlateDBs can get
aggregated stats for the cache.