write manifests more safely by verifying the expected next id (#496)
This patch refactors the manifest store by changing the way that
manifests are updated. To update a manifest you now call prepare_dirty,
which returns an instance of DirtyManifest. DirtyManifest contains the
manifest contents, along with its id, and the id that it will be written
with (next_id). You then make changes to the dirty manifest, and pass it
to update_manifest. update_manifest verifies that next_id matches the
expected next id before writing out the dirty manifest.

DbState/CompactorState now work by storing a DirtyManifest with the
local changes, rather than just CoreDbState, and the db/compactor use
this dirty manifest when updating the manifest.

This change confers two benefits:

Firstly, and the primary goal of this patch, it makes manifest updates
safer by reducing the likelihood of accidentally writing a stale
manifest version. Previously, the following sequence was possible:
1. Writer/Compactor merges the current manifest into its local state
(DbState /CompactorState)
2. Writer/Compactor updates the manifest without going through the local
state. For example, maybe it creates a checkpoint directly in its
FenceableManifest.
3. Writer/Compactor writes the manifest from local state to
FenceableManifest. The write will go through because FeneableManifest is
fully up-to-date, though the local state is not. After this patch, if
the local state's dirty manifest is not fully up-to-date with
FenceableManifest/StoredManifest, then the update will fail.

Secondly, DirtyManifest has information about the id and next id. This
means we can do checkpoint management from outside
StoredManifest/FenceableManifest, which feels out of place currently.
This change is not implemented in this patch however