Draft RFC with guidance on public API errors (#471)
This is an RFC about the usage of errors in the public API . It is based
on some thoughts I shared previously:
https://github.com/slatedb/slatedb/issues/454. I've begun with an
intentionally small set of public errors. If there are any obvious
additions, we can add them here, or we can wait until they emerge
through refactoring.

Feedback is much appreciated.

Fixes #454

---------

Co-authored-by: Chris Riccomini <criccomini@users.noreply.github.com>