Factor out WAL replay utility (#500)
In https://github.com/slatedb/slatedb/pull/498, the reader will need to
replay WALs similar to the way that a writer does. This patch factors
the core replay logic out of the writer to make it possible to reuse it
in the reader. This also makes it easier to write tests for WAL replay
and I've added some new test cases.