Add version bytes for SST and manifest byte formats (#510)
This PR adds very rudimentary version number handling for SSTs and
manifests. My main goal is simply to get the version bytes in the right
spot, so we can implement proper version handling going forward. The
most important part of this PR are the byte sizes and locations in the
byte formats.

Changes:

- Adds a u16 to the _end_ of an SST's on-disk byte format
- Adds a u16 to the _beginning_ of a manifest's on-disk byte format
- Adds version validation logic in `FlatBufferManifestCodec::decode`
- Adds version validation logic in `SsTableFormat::read_info`
- Adds an `InvalidVersion` error
- Adds tests for validation logic

A couple of notes/oddities:

- I opted to put the encode/decode logic inside the manifest's flat
buffer codec because we don't currently have an intermediate builder for
it the way that we do for SSTs (using `EncodedSsTableBuilder`).
- On the SST side, I opted to validate the version in `read_info` since
you can't do anything without reading the SST info for an SST first.
- The SST bytes are at the end while they're at the beginning for the
Manifest. This felt most intuitive to me, but if everyone is grossed out
by it, I can put the version bytes at the heat of the SST or tail of the
manifest.
- The current decoders simply validate the versions are identical. We
only have a single version right now, so this is fine. The work to
support cross-version decoding is left for a later PR.

The changes are a first step toward #446.