fix: handle repeated keys during sst index search
This patch fixes the SstIterator's index search path to
correctly handle SSTs where multiple blocks contain the
same starting key (which is now possible in l0 since we
store multiple key versions in the memtable). The fix
is to not terminate the binary search of the block metas
when a match is found - but rather to keep looking for a
lower-indexed block.

fixes #517