fix block filtering when creating sst iterator (#522)
This patch fixes two issues with selecting the start and end blocks for
an sst iterator:
- for the start block, in the case where iteration starts from the same
key as the first key in a block B, we were selecting B as the start
block. Instead, if B > 0 we should start from B-1 because block B-1 may
end with the specified key.
- for the last block, we were selecting the block after the start block.
After fixing the above bug, in the case where a key first appears as the
first key in some block B, we wind up excluding B because the search
spans [B-1,B).

This patch makes a few changes to fix these problems:
- It adds a trait called RangePartitionedKeySpace that defines a key
space that is split up into a list of partitions, each of which has a
first key. We need this trait because we can't treat the block meta as a
slice to get the nice binary search fns. We can't treat it as a slice
because it uses a custom flatbuffer vector type.
- It implements RangePartitionedKeySpace for SsTableIndex.
- It provides the following helper fns that operate on
RangePartitionedKeySpace
- first_partition_including_key: returns the first partition that could
include data with a given key
- last_partition_including_key: returns the last partition that could
include data with a given key
- SST Iterator now uses these helper fns to bound the iteration space.

In a later patch we can also implement RangePartitionedKeySpace for
SortedRun so we can use the same search fns to triangulate a keys
position in the ssts in a sorted run.

Fixes #517