Add Serialize/Deserialize impls for CachedKey/CachedEntry (#525)
This patch implements Serialize/Deserialize for CachedKey/CachedEntry so
that we can then support db cache impls that write cache entries to
local disk. The impls work by first converting CachedKey and CachedEntry
to/from types that #[derive] Serialize/Deserialize. This way the
in-memory representations are decoupled from the on-disk format.