add a dbcache impl that uses the foyer hybrid cache (#526)
This patch adds an implementation of `DbCache` that uses a foyer hybrid
cache to store cached blocks on-disk as well as in-memory.
`FoyerHybridCache` can fail gets, so this patch also changes `DbCache`
to support returning errors. Currently errors are handled as follows:
- tablestore just ignores errors and treats them as cache misses
- `DbCacheWrapper` tracks a counter of errors, and logs them. The log is
rate-limited to once per second (so as not to spam logs on non-transient
errors like out-of-disk).