change read level to durability filter (#529)
This patch changes the current ReadLevel and ReadOptions.read_level to
DurabilityLevel and ReadOptions.durability_filter as described in RFC-8.
It also changes the read path when wal is disabled to respect the
durability filter by only reading from the memtable if the
durability_filter is set to Memory. Previously we were disregarding the
read level when wal was disabled.