refactor: let MergeIterator accepts Box<dyn KeyValueIterator>; merge SeekToKey into KeyValueIterator (#546)
currently there's a limitation in `MergeIterator`: the children iterator
passed to it has to be the same type.

this pr makes KeyValueIterator able to be used as a trait object, so we
can pass different implementations of `KeyValueIterator` into it.

Rust type system does not allow things like `Box<dyn KeyValueIterator +
SeekToKey>`, so this pr has to merge the trait `SeekToKey` into
`KeyValueIterator` together.