fix: fix build error when using dbiterator behind a Box<dyn>
Before this patch, if you tried to use DbIterator behind a Box<dyn>
you would get build errors complaining about the lifetime associated
with the closure used to generate the memtable iterators. I think this
is because of the extra constraints introduced by async_trait. To
fix this, we collect the iterators into a Vec before passing them
to MergeIterator. Also includes a test that replicates the error