cache l0 SSTs on write (#528)
This patch changes the memtable flush path to write L0 sst data to the
db cache when a memtable is flushed. The idea is that this data was
recently written and so is more likely to be read than other data in the
cache.

This change required a few supporting changes from
EncodedSsTable(Builder):
- Rather than storing a list of raw block bytes, the builder now stores
a list of EncodedSsTableBlock which contains the encoded + compressed
bytes, the decoded block, and its offset so that it can be easily
inserted into the cache.
- The implementation of the builder is refactored a bit to support
maintianing this list. finish_block now calls to a new fn called
prepare_block which constructs EncodedSsTableBlock, and finish_block is
in charge of setting the builder state to reflect the completion of the
current block and start of the next block.
- EncodedSsTable now has fns for writing all the bytes into a buffer,
which the tests and tablestore use rather than re-implementing this
logic everywhere.