[bug] db_options.flush_interval=None broken (#548)
## the problem
I noticed that setting `flush_interval` as `None` currently does not
work. If it's set as `None`, currently the flush task just stops:

https://github.com/slatedb/slatedb/blob/9e4a80b37fd91cd29b7a485712ea3cb2ba7388f8/src/flush.rs#L101-L104
However this task not only performs periodic flushes, it also handles
manual flush requests (which don't work in this case):

https://github.com/slatedb/slatedb/blob/9e4a80b37fd91cd29b7a485712ea3cb2ba7388f8/src/flush.rs#L128-L129

## fix
When `flush_interval` is `None`, change the behavior to start the flush
task- just without the periodic ticking.

Co-authored-by: sush-abc <sush-abc@users.noreply.github.com>