fix(compaction): filter tombstones out when compacting into the last sorted run (#307)
This PR will close #285

Unfortunately I didn't come up with a good way to add unit tests,
because tombstones behave the same when accessed via the `Get` operation
regardless of whether they're physically removed or not.

I followed @rodesai suggestion of letting the orchestrator decide
whether the dest sorted run is the last sorted run or not.

---------

Co-authored-by: Almog Gavra <almog.gavra@gmail.com>