Verify that compression works without panics (#561)
I added a test based on the bug report from #555. The test successfully
replicated the issue on tag v0.6.1, but appears to have already been
fixed on main. I didn't dig into why. Perhaps it's related to #548. In
any case, I want to include a test just to make sure we don't hit the
issue again in the future.

Fixes #555