RFC: Settings vs components (#559)
Hi folks 👋 

I was reading some of the latest changes in SlateDB and I wanted to
bring forward a proposal to be more intentional about how to configure
the engine.

When I implemented the [ability to load settings from plain text
files](https://github.com/slatedb/slatedb/pull/270) I noticed that the
`DbOptions` struct has a mix of settings that a user can configure via
text, and components that the database uses for different things, like
the `gc_runtime`.

#533 adds a new component that instead of being in `DBOptions`, is
passed to the `open_with_opts` functions directly.

It's starting to sound to me like there is no clear pattern for adding
settings and configurable components that makes it easy to keep future
and past compatibility stable.

I wrote a proposal to improve this situation with the goal of making it
easier in the future to decide how configurable components are added
while keeping backwards compatibility.

Let me know what you think.