refactor: Extract bencher, cli, and bytes-generator to independent crates (#554)
This change extract three submodules into independent crates to isolate
them from the core of SlateDB. Building these crates becomes optional.

The drawback of this change is that many of the internal core structs
need to be public for the bencher to use them. It might be worth in the
future to consider a different abstraction for those structs, but it
should not be too concerning to make them public.

I guess you can assign #460 to me. However, I think that issue is to
broad to tackle, and this is probably the MVP I'd consider good enough
for now. These changes should create minimal disruption in other PRs,
bigger changes could be worse for PRs already open, and I'd advocate for
not making the lives of the authors of those PRs more difficult that it
needs to be.