refactor(wal): skip entries with seq that lower than last_l0_seq on replaying (#569)
this PR is also groundwork for #519.

as outlined in RFC-008
(https://github.com/slatedb/slatedb/blob/main/rfcs/0008-synchronous-commit.md),
the way we write WALs might change, the entries of WAL files won't
necessarily line up perfectly with IMM/L0 anymore.

right now, the flow is:

- writes are appended to the WAL first.
- when the WAL buffer hits its limit, it's frozen into an immutable
buffer.
- this frozen buffer is flushed to object storage.
- the writes from this buffer are applied to the MemTable during the
freeze.
- when MemTable reaches its limit, it's frozen into an IMM, and flushed
to L0.

```
--------------------------
| WAL Buf 1 |            |
-------------  L0 SST 1  |
| WAL Buf 2 |            |
--------------------------
```

in the current design, when a MemTable is frozen into an IMM and stored
in L0, its `last_seq` will exactly align with the `last_seq` of its last
WAL SST.

this means during recovery, we can just start recovering from L0's
`last_wal_id` + 1 and don't need to worry about the `last_seq`.

but after RFC-008, we won't wait for the WAL to fill up before applying
writes to the MemTable.

instead, when a user writes, we'll write to the WAL first and then
immediately apply that write to the MemTable.

the WAL files will be saved to storage on their own schedule and won't
necessarily align with the contents of the IMM/L0 anymore.

when a MemTable is frozen into an IMM and persisted to L0, its
corresponding WAL entries might not have been persisted yet. in this
case, it's still good to record the most recent persisted WAL ID at this
moment, to help us navigate the starting WAL ID during recovery.

the difference is, we'll need to skip the WAL entries that have already
been applied in L0 with filtering out the WAL entires whose `seq` is
smaller than the last L0's `last_seq`.

this pr adds the following changes:

1. rename `wal_id_last_compacted` to `last_l0_recent_flushed_wal_id`,
let this recent flushed wal id be considered as an attribute of the
`last_l0`, and "recent" is a weaker assumption than "last". it can be
considered as a hint to let us find out where to start the replay, the
correctness of replay posisition is guaranateed by `l0_last_seq`.
2. filter out the entries whose `seq` is smaller than `last_l0_seq`
during replay.

this pr is expected to not change any of the current behaviour of
replaying/recovery. all the behaviour changes will be enabled after #519