fix(gc): Reimplement the garbage collector without crossbeam.
Introduce a simplified version of the GarbageCollector that doesn't use crossbeam::channels for concurrency control.

This new implementation also removes the state machine that made the code more complex in favor of purposely defined functions. This fixes some fleaky tests that could crash because the GC thread terminated ahead of time and made the database enter an error state.

This change also asumes that the admin tools don't run in the same process than the actual database, hence there is no need of a fully dedicated system thread to run the garbage collector, it can run in a Tokio task. The caller can decide to run the task in a separated system thread if it's necessary for the task at hand.