Fix flaky `test_compactor_compacts_l0` (#597)
It appears we have a flaky compactor test. It's reported in #595. I was
able to reproduce the test by setting the `compactor_options`
`poll_interval` to `1ms`. This resulted in a lot of random failures.

There are a ton of problems with this test.

- It assumes there will be one SR
- It assumes there will be one SST in that SR
- It assumes all data will make it into that one SST

None of these assumptions are true.

To fix this, I've:

1. Updated `await_compaction` to wait for all data to make it into at
least L1+.
2. Updated the test to use a size tiered scheduler that slurps up pretty
much all L0 SSTs whenever possible. Without this, the test would see
backpressure and/or conflicts.
3. Updated the test to verify that all keys are in L1+.
4. Updated the test to verify the db returns the data as expected
through `get()`. This was a `// todo`.

Fixes #595