Add `LogicalClock` and `SystemClock` clocks (#601)
This PR does the following:

- Creates `LogicalTime` and `SystemTime` types, which return i64 and
`SystemTime`, respectively.
- Creates a `DefaultSystemTime` that replaced `TokioClock`. This clock
gives us normal SystemTimes by default, but allows us to pause the tokio
runtime to take control of the clock for DST.
- Creates a `DefaultLogicalTime` that simply uses `DefaultSystemTime`
inside.
- Replaces all usage of `Clock` with `LogicalClock`.
- Adds `with_system_clock` and changes `with_clock` to
`with_logical_clock`

This PR is a less ambitious version of @calavera's PR (#586). Unlike
#586, this PR does not...

1. Attempt to replace all usage of system time (SystemTime::now() and
tokio time) with a SystemClock.
2. Attempt to move MonotonicClock. Those two steps are left for future
PRs.

These two steps will be left for a follow-on PR.