Add `Admin` and `AdminBuilder` (#610)
I found that our public facing static functions are a real pain to work
with while working on #606 and #609. I initially thought using a
`DbContext` that held just normally static stuff (rng, clock, etc) would
be helpful. After working through it, I felt the right thing to do was
get rid of `DbContext` and implement proper builders for user-facing
stuff. That includes:

- admin.rs
- compactor.rs (and related compactor stuff)
- garbage_collector.rs

Each of these builders will allow us to do `_with_system_clock` and
`_with_logical_clock` if/when needed. As we add an RNG and filesystem
for DST, they'll also have `_with_*` methods for the appropriate
builders.

Internally, we'll pass actual `impl SystemClock`, `impl LogicalClock`,
and so on to the structs and functions that need them.

This PR currently:

- implements `Admin` and`AdminBuilder` to show what the improved UX will
be.
- moves `delete_objects_with_prefix` to the bencher `main.rs` file since
it's only used there and has nothing to do with SlateDB.
- updates main.rs files to use the new `Admin` client.
- moves `refresh_checkpoint` and `delete_checkpoint` out of `Db`
(`checkpoint.rs`) and into `Admin.